refactor: clean up mess in test files (#1314)
The "toolbar-buttons" test was actually just another "toolbar-add" test,
so I moved its contents into the "toolbar-add" test file.

The "widget-interaction-point" test was actually a "widget-position"
test, and in the last commit I folded the latter into "toolbar-buttons",
so I renamed that one the "toolbar-buttons" test.
